AES keys stored in database - these will eventually be cycled.
As this is an RPi project, I am taking into consideration how much security is actually needed - as the RPi will be locked down as much as possible, I am comfortable with allowing the private keys to be kept in the database. In a perfect world, one would have a HSM to manage keys, but this is not the case.